Make PNG support optional
