+++ src/element.rs
@@ -1,6 +1,6 @@
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::cmp;
-use std::io::{self, Cursor, Error, ErrorKind, Read, Write};
+use std::io::{self, Error, ErrorKind, Read, Write};
 
 use super::icontype::{Encoding, IconType, OSType};
 use super::image::{Image, PixelFormat};
@@ -9,6 +9,7 @@ use super::image::{Image, PixelFormat};
 const ICON_ELEMENT_HEADER_LENGTH: u32 = 8;
 
 /// The first twelve bytes of a JPEG 2000 file are always this:
+#[cfg(feature = "pngio")]
 const JPEG_2000_FILE_MAGIC_NUMBER: [u8; 12] =
     [0x00, 0x00, 0x00, 0x0C, 0x6A, 0x50, 0x20, 0x20, 0x0D, 0x0A, 0x87, 0x0A];
 
@@ -60,10 +61,13 @@ impl IconElement {
         }
         let mut data: Vec<u8>;
         match icon_type.encoding() {
+            #[cfg(feature = "pngio")]
             Encoding::JP2PNG => {
                 data = Vec::new();
-                try!(image.write_png(&mut data));
+                image.write_png(&mut data)?;
             }
+            #[cfg(not(feature = "pngio"))]
+            Encoding::JP2PNG => unimplemented!(),
             Encoding::RLE24 => {
                 let num_pixels = (width * height) as usize;
                 match image.pixel_format() {
@@ -110,13 +114,14 @@ impl IconElement {
         let width = icon_type.pixel_width();
         let height = icon_type.pixel_width();
         match icon_type.encoding() {
+            #[cfg(feature = "pngio")]
             Encoding::JP2PNG => {
                 if self.data.starts_with(&JPEG_2000_FILE_MAGIC_NUMBER) {
                     let msg = "element to be decoded contains JPEG 2000 \
                                data, which is not yet supported";
                     return Err(Error::new(ErrorKind::InvalidInput, msg));
                 }
-                let image = try!(Image::read_png(Cursor::new(&self.data)));
+                let image = try!(Image::read_png(io::Cursor::new(&self.data)));
                 if image.width() != width || image.height() != height {
                     let msg = format!("decoded PNG has wrong dimensions \
                                        ({}x{} instead of {}x{})",
@@ -128,6 +133,8 @@ impl IconElement {
                 }
                 Ok(image)
             }
+            #[cfg(not(feature = "pngio"))]
+            Encoding::JP2PNG => unimplemented!(),
             Encoding::RLE24 => {
                 let mut image = Image::new(PixelFormat::RGB, width, height);
                 try!(decode_rle(&self.data, 3, image.data_mut()));
