+++ src/image.rs
@@ -1,7 +1,5 @@
-use png;
-use png::HasParameters;
 use std;
-use std::io::{self, Read, Write};
+use std::io;
 
 /// A decoded icon image.
 ///
@@ -16,10 +14,10 @@ use std::io::{self, Read, Write};
 /// bottom-right pixel comes last).
 #[derive(Clone)]
 pub struct Image {
-    format: PixelFormat,
-    width: u32,
-    height: u32,
-    data: Box<[u8]>,
+    pub(crate) format: PixelFormat,
+    pub(crate) width: u32,
+    pub(crate) height: u32,
+    pub(crate) data: Box<[u8]>,
 }
 
 impl Image {
@@ -150,60 +148,6 @@ impl Image {
             height: self.height,
             data: new_data,
         }
-    }
-
-    /// Reads an image from a PNG file.
-    pub fn read_png<R: Read>(input: R) -> io::Result<Image> {
-        let decoder = png::Decoder::new(input);
-        let (info, mut reader) = try!(decoder.read_info());
-        let pixel_format = match info.color_type {
-            png::ColorType::RGBA => PixelFormat::RGBA,
-            png::ColorType::RGB => PixelFormat::RGB,
-            png::ColorType::GrayscaleAlpha => PixelFormat::GrayAlpha,
-            png::ColorType::Grayscale => PixelFormat::Gray,
-            _ => {
-                // TODO: Support other color types.
-                return Err(io::Error::new(io::ErrorKind::InvalidData,
-                                          format!("unsupported PNG color \
-                                                   type: {:?}",
-                                                  info.color_type)));
-            }
-        };
-        if info.bit_depth != png::BitDepth::Eight {
-            // TODO: Support other bit depths.
-            return Err(io::Error::new(io::ErrorKind::InvalidData,
-                                      format!("unsupported PNG bit depth: \
-                                               {:?}",
-                                              info.bit_depth)));
-
-        }
-        let mut image = Image::new(pixel_format, info.width, info.height);
-        assert_eq!(image.data().len(), info.buffer_size());
-        try!(reader.next_frame(image.data_mut()));
-        Ok(image)
-    }
-
-    /// Writes the image to a PNG file.
-    pub fn write_png<W: Write>(&self, output: W) -> io::Result<()> {
-        let color_type = match self.format {
-            PixelFormat::RGBA => png::ColorType::RGBA,
-            PixelFormat::RGB => png::ColorType::RGB,
-            PixelFormat::GrayAlpha => png::ColorType::GrayscaleAlpha,
-            PixelFormat::Gray => png::ColorType::Grayscale,
-            PixelFormat::Alpha => {
-                return self.convert_to(PixelFormat::GrayAlpha)
-                    .write_png(output);
-            }
-        };
-        let mut encoder = png::Encoder::new(output, self.width, self.height);
-        encoder.set(color_type).set(png::BitDepth::Eight);
-        let mut writer = try!(encoder.write_header());
-        writer.write_image_data(&self.data).map_err(|err| match err {
-            png::EncodingError::IoError(err) => err,
-            png::EncodingError::Format(msg) => {
-                io::Error::new(io::ErrorKind::InvalidData, msg.into_owned())
-            }
-        })
     }
 }
 
@@ -717,6 +661,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "pngio")]
     fn write_grayscale_png() {
         let gray_data: Vec<u8> = vec![63, 127, 191, 255];
         let mut image = Image::new(PixelFormat::Gray, 2, 2);
@@ -733,6 +678,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "pngio")]
     fn write_rgb_png() {
         let rgb_data: Vec<u8> = vec![255, 0, 0, 0, 255, 0, 0, 0, 255, 127,
                                      127, 127];
@@ -751,6 +697,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "pngio")]
     fn write_rgba_png() {
         let rgba_data: Vec<u8> = vec![255, 0, 0, 63, 0, 255, 0, 127, 0, 0,
                                       255, 191, 127, 127, 127, 255];
@@ -769,6 +716,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "pngio")]
     fn read_rgba_png() {
         let png: Vec<u8> =
             vec![137, 80, 78, 71, 13, 10, 26, 10, 0, 0, 0, 13, 73, 72, 68,
@@ -788,6 +736,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "pngio")]
     fn png_round_trip() {
         let rgba_data: Vec<u8> = vec![127, 0, 0, 63, 0, 191, 0, 127, 0, 0,
                                       255, 191, 127, 127, 127, 255];
